+++ README.md
@@ -6,7 +6,7 @@ then 停车成功返回一个凭证
 
 given 一个停车场有2个空车位  
 when 依次停入2辆汽车  
-then 停车成功返回2个凭证  
+then 停车成功返回2个不同凭证  
 
 given 一个停车场有一个空车位  
 when 依次停入2辆汽车  
@@ -101,3 +101,21 @@ then 取车成功
 given 我有一张该停车场的取车凭证 
 when 停车小弟连续两次取出汽车  
 then 第一次取车成功，第二次取车失败  
+
+## parking car manager
+## Tasks
+given 1个停车管理员管理1个停车小弟，这个小弟有一个停车场并且有一个空位
+when 有1辆车需要停
+then 停车小弟停车成功返回一个凭证
+
+given 1个停车管理员管理2个停车小弟，小弟个有一个停车场并且有一个空位
+when 有2辆车需要停
+then 两个小弟分别停车成功，均成功返回凭证
+
+given 1个停车管理员管理1个停车小弟，这个小弟有一个停车场并且没有空位
+when 有1辆车需要停
+then 停车管理员自己停车成功
+
+given 我有一辆车需要取出来，用有效的凭证去找管理员
+when 取车
+then 管理员取车成功
