--- src/main/java/parkinglot/Car.java
@@ -1,4 +1,0 @@
-package parkinglot;
-
-public class Car {
-}
