--- src/main/java/parkinglot/ParkingLotReceipt.java
@@ -1,11 +1,0 @@
-package parkinglot;
-
-import java.util.UUID;
-
-public class ParkingLotReceipt {
-    private final String receipt;
-
-    public ParkingLotReceipt() {
-        receipt = UUID.randomUUID().toString();
-    }
-}
