+++ src/main/java/parkinglot/beta/Car.java
@@ -1,0 +1,4 @@
+package parkinglot.beta;
+
+public class Car {
+}
