+++ src/main/java/parkinglot/beta/Receipt.java
@@ -1,0 +1,4 @@
+package parkinglot.beta;
+
+public class Receipt {
+}
