+++ src/main/java/parkinglot/beta/exceptions/NoSpaceInParkingLotException.java
@@ -1,0 +1,7 @@
+package parkinglot.beta.exceptions;
+
+public class NoSpaceInParkingLotException extends Exception {
+    public NoSpaceInParkingLotException(String message) {
+        super(message);
+    }
+}
